+++ src/components/CustomResources/CustomResourcesByGroup.js
@@ -1,6 +1,5 @@
-import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { Link } from 'shared/components/Link/Link';
+import { ExternalLink } from 'shared/components/ExternalLink/ExternalLink';
 import { Link as RRLink } from 'react-router-dom';
 import pluralize from 'pluralize';
 import { GroupingListPage } from './GroupingListPage';
@@ -11,10 +10,7 @@ export default function CustomResourcesByGroup({ enableColumnLayout }) {
   const { namespace, clusterUrl, scopedUrl } = useUrl();
   const description = (
     <Trans i18nKey="custom-resources.description">
-      <Link
-        className="bsl-link"
-        url="https://kubernetes.io/docs/concepts/extend-kubernetes/api-extension/custom-resources/"
-      />
+      <ExternalLink url="https://kubernetes.io/docs/concepts/extend-kubernetes/api-extension/custom-resources/" />
       <RRLink
         className="bsl-link"
         to={clusterUrl(`customresourcedefinitions`)}
